Fixes path to artifact.
The way path was written was forcing the 'output' directory itself to be
included in the zip file (while we only care about its root/etc).

The problem was because it was edited to 'output**' instead of
'output/**', but it is enough to specify either 'output' or
'output/' to use the whole directory as root (but not include it).